Add support for a variety of Kubernetes objects
K8s logic has been moved to an internal package, so the code is more clean.
Additional object kinds are added by providing an implementation of the "applier" interface,
and then registering the object kind in the "supportedObjectKinds" map.

This commit adds support for batch/v1beta1.CronJob as well as the
already supported apps/v1.Deployment